feat(pkt-io): impl Clone for PktIo NF
Each PktIO instance can have 2 queues for injection and punting.
In some cases, e.g. when having multiple pipelines, we may want
the PktIO instances in each to appear as one, sharing the same
inject and punt queues. This is convenient, for instance, to
inject locally generated packets, without needing to specify to
which of the N identical pipelines the packet should be injected
into. Implementing Clone for PktIo allows multiple instances
in distinct pipelines to appear as one.

Performance note: in principle, the PktIo uses lockless queues.
If having a single shared queue across multiple workers is
an issue, the IO manager can be extended to use multiple queues
and round-robin on them.

Signed-off-by: Fredi Raspall <fredi@githedgehog.com>