feat(net): extend warning on drop for sourced pkts
When dropping a packet (its metadata), we log an error if its
done state has not been set, in order to detect logical errors in
a pipeline where a packet would not be processed. We need to extend
the criteria because a pipeline may be arranged such that its
last stage injected packets, which no other stage would process.
Since locally originated packets are marked as "sourced", don't
complain if we drop such packets without a done reason being set.

Signed-off-by: Fredi Raspall <fredi@githedgehog.com>