bump(rtnetlink): rebase onto upstream changes
Quite a few bug fixes have landed in rtnetlink upstream.

Also, it is best that we be as close as possible to an actual release of rtnetlink before we launch.

If nothing else we should see significantly reduction in dependency bloat, as rtnetlink bumped to thiserror 2 (among
other) dependencies which we now share.

Signed-off-by: Daniel Noland <daniel@githedgehog.com>