fix(nat): Reuse flow table from flow table lookup stage
When we added the stateful NAT stage to dataplane's pipeline, we made it
create a new flow table, from scratch, rather than reusing the existing
one from the flow table lookup stage. This is bad: this means that we
create NAT sessions in a table different than the one we use for
lookups, so we never add entries to the lookup table and never find
existing sessions. For packets going out, this means we recreate a
session every time; for packets coming back, we deny session creation
and the packets are dropped.

Change stateful NAT setup in order to reuse the flow table from the
lookup stage.

Fixes: a09de33a3d78 ("feat(dataplane): Add stateful NAT stage to pipeline")
Signed-off-by: Quentin Monnet <qmo@qmon.net>