fix(nat): Use dest port for ICMP identifier if source port is empty
Our support for ICMP in stateful NAT is a bit hacky, and at the moment
we reuse the "source port" field (for TCP/UDP) in order to store the
ICMP Echo messages identifier to translate in the flow table. This works
well enough, except that for return packets, we're not necessarily
doing source NAT, and the source port may be empty as a result. In that
case, use the destination port instead. Make sure to only do so when we
don't have a source port, to avoid conflicting values for the identifier
(the allocator currently allocates distinct values for source and
destination ports; it needs some rework to better support ICMP, but this
will be for future work).

With this change, stateful NAT on the return path in case of
unidirectional NAT should work. Add a test to ensure that it does.

Fixes: d51311bd12ce ("feat(nat): Add logic for processing ICMP in stateful NAT")
Signed-off-by: Quentin Monnet <qmo@qmon.net>