fix(nat): Allow un-NATed traffic from VpcExpose without NAT
With stateful NAT, we reject session creation when we are not able to
find source NAT information. This is to prevent connections coming "from
the outside" of a VPC with stateful NAT configured to open new sessions,
sessions can only be initiated "from the inside".

However, we enforce this rule on all connections, ... whether or not
they pertain to a VpcExpose that has NAT configured. As a consequence,
traffic associated to a peering that does not rely on NAT would be drop,
because we cannot find source NAT information (obviously).

To address this issue, we introduce stateless "exemption tables"
containing valid prefixes from VpcExpose objects that do not use
stateful NAT. Traffic associated with these prefixes can just go through
without being NAT-ed. Traffic associated with VpcExpose objects using
stateful NAT will be NAT-ed, as expected. Traffic associated with no
exposed IP will be dropped. We add a test with no NAT configured, to be
sure.

Fixes: fb45c063b418 ("feat(nat): Reject stateful NAT connection if there is no source NAT rule")
Signed-off-by: Quentin Monnet <qmo@qmon.net>