fix(nat): Fix "NAT-exempt packet" lookup for stateful NAT
We introduced a lookup to check whether packets would be exempt of
stateful NAT. We added the lookup before checking for existing sessions
for the packet. This works for packets going forward and receiving
source NAT, because they fail the exemption lookup and go through the
session lookup. However, when packets come back, they may be associated
with a source VpcExpose that does not require NAT (so might be
considered as "exempt") but may still require destination NAT if they're
part of an existing session! In that case, passing the exemption check
will prevent them from getting destination NAT, and will result in these
packets being dropped by the router.

To fix this, move the exemption lookup after the flow table lookup.

Adjust the corresponding test that had empty "IPs" for the second
VpcExpose, and as such was not triggering the issue.

Fixes: a08366512b49 ("fix(nat): Allow un-NATed traffic from VpcExpose without NAT")
Signed-off-by: Quentin Monnet <qmo@qmon.net>