fix(nat): Disable stateful + stateless NAT
Re-introduce the drop of the NAT flag at the end of stateless NAT. This
commit differs from cb9f242c0e5a ("feat(nat): unmark nat flag after
successful nat"), because we only unmark the packet when stateless NAT
has changed the packet.

The issue with stateless + stateful is that the stateful NAT stage needs
to know whether or not the packet requires NAT, and for doing so it
currently runs a "NAT-exempt" check using the source IP for the packet.
But this source IP is validated against the list of IPs of the VpcExpose
blocks and does not account for address changes from stateless NAT.

There are a number of better, but more involved possibilities to address
this. Here we go with the simplest fix.

Fixes: 01b20ce08147 ("Revert "feat(nat): unmark nat flag after successful nat"")
Signed-off-by: Quentin Monnet <qmo@qmon.net>