feat(pkt-meta): Do not return dst VPC discriminant on conflicts
We have a pipeline stage that returns the destination VPC discriminant
for a packet, based on its source VPC discriminant and its destination
address. This works well, except in case of conflict: If VpcExpose
blocks from several VPCs expose the same IPs to a given VPC, then the
source VPC and destination address may not be enough to tell what the
destination VPC is.

The table for the lookups would not deal with this situation well: when
we built the table, we would simply overwrite the entry value (the
destination VPC discriminant) with the latest discriminant found in the
VPC tables for a given prefix. Instead, make the value an
Option<VpcDiscriminant>, so that we can return None in case of conflict,
even though we have a corresponding prefix in the table.

Also update the lookup stage to account for this case, and add a unit
test to validate the construction of the table.

Signed-off-by: Quentin Monnet <qmo@qmon.net>