feat(nat): Make dest VPC lookup table writer available to stateful NAT
In order to add support for shared IP addresses exposed to a given VPCs
by multiple VPCs, we must update the tables for destination VPC
discriminant lookup from the stateful NAT pipeline stage. To that end,
pass a (protected) writer for the table to the stateful NAT stage. It
will be used in a follow-up commit.

TO DO:
 - Rework VPC discriminant lookup tables to use something else than
   left-right.
 - Get rid of the Mutex.

Signed-off-by: Quentin Monnet <qmo@qmon.net>