feat(nat): Update destination VPC discriminant lookup tables
In order to add support for shared IP addresses exposed to a given VPCs
by multiple VPCs, we must update the tables for destination VPC
discriminant lookup from the stateful NAT pipeline stage, so that we're
able to tell what VPC a packet is heading to, based on information from
existing sessions. We insert a new entry in the table just after
creating a new reverse session.

Add the relevant tests.

TO DO:
 - Clean up entries when dropping sessions
 - Solve the DoS case (see tests).

Signed-off-by: Quentin Monnet <qmo@qmon.net>