fix(nat): Fix NatPort conversions
We recently changed the NatPort type to make it an enum so it could
represent either a potentially-null identifier, or a non-null port.
Following that update, we also need to make sure that we cannot convert
from any enum variant into a TcpPort or UdpPort, for example: we can
only do so if we handle a NatPort::Port() variant. Let's fix the
relevant methods, even though this bug does not seem to currently affect
any of the NAT logic.

Fixes: 51f2e5668635 ("feat(nat): Change NatPort to allow (possibly-null) identifiers")
Signed-off-by: Quentin Monnet <qmo@qmon.net>