feat(nat): Move stateless NAT prefix size validation back to "config"
In commit 28c87043f59a ("feat(nat): Move stateless NAT prefix size
validation out of "config""), we moved prefix size validation for
stateless NAT out of "config", and into the "nat" crate instead. This
was because we had no way to tell whether the VpcExpose block uses
stateless NAT, and whether this check was relevant, so we had to move it
out of the way to allow stateful NAT to work with IP sets of different
sizes.

Now that we do have a way to tell, via the presence and configuration
variant of VpcExposeNat objects, it's trivial to apply the check only to
VpcExpose that actually rely on stateless NAT. Let's move validation
back to the "config" crate, along with the other validation steps. This
commit is more or less a rebased revert of the aforementioned one,
except that we filter exposes with self.has_stateless_nat() for running
the check.

Signed-off-by: Quentin Monnet <qmo@qmon.net>