feat(dataplane): Add a force_send method
In a tokio runtime, things used across an await need to be Send
because the future created can be continued in a different thread.
However, we use a current_thread runtime which won't do this
so this commit adds a method to silence the incorrect error from
the Rust compiler when using non-Send objects across await, like
the pipeline itself.

Signed-off-by: Manish Vachharajani <manish@githedgehog.com>