feat(dataplane): PACKET_FANOUT based kernel driver
The prior kernel drivers used a single AF_PACKET socket and
a read thread that serially read and spread packets to workers.
This is inefficient and creates a single thread bottleneck for
scaling up workers.

This driver uses the kernel's PACKET_FANOUT option that lets
the kernel fanout packets to multiple AF_PACKET sockets. This
commit uses PACKET_FANOUT_HASH but we should also experiment
with setting PACKET_FANOUT_QM to use the hardware RSS queues.

Signed-off-by: Manish Vachharajani <manish@githedgehog.com>