feat(pkt-meta): Move "unroutable" drop from dst VPC to flow lookup
On failure to determine the destination VPC discriminant for a packet at
the destination VPC lookup stage, we would drop the packet as
"unroutable".

There is a case when we want to do differently, though: when IP
addresses from overlapping ranges are exposed by multiple VPCs to a
given VPC, the destination VPC lookup may fail, but a flow table lookup
may instead provide us with the discriminant based on existing flow
states. To support that, we change the destination VPC lookup stage to
allow packets when we are not able to uniquely tell what VPC the packet
is addressed to.

Given that we remove this drop in the destination VPC lookup stage,
however, we must make sure we compensate for it in the flow lookup
stage: make sure to drop the packet in that stage if we still ignore the
destination VPC after the flow lookup.

Why not add an additional flow table lookup as part of the destination
VPC lookup stage? The logic would be correct, but it would create a
duplicate lookup (the flow table lookup would still come right after in
the pipeline), and we don't want that. Furthermore, the flow table
lookup already has all the infrastructure to retrieve and attach the
destination VPC discriminant, if found, to the packet: it's the natural
stage where to do it.

Also add a name attribute to the flow table stage, so we can print it in
the logs.

Signed-off-by: Quentin Monnet <qmo@qmon.net>