feat(nat): Write destination VPC discriminant to flow table values
So far, the destination VPC discriminant for a packet has been part of
the flow key, not the flow value. This was based on the assumption that
it's always available when we do stateful NAT, in other words, that
we're always able to tell, during the destination VPC lookup stage, what
the destination VPC for a given packet is, based on its source VPC and
its destination address.

This assumption no longer holds in the case where multiple VPCs expose
the same addresses to a given VPC: for example when VPC-1 and VPC-3 both
expose 10.0.0.0/24 to VPC-2 with NAT, the requests to VPC-2 are easy to
associate to the destination VPC, but the replies from VPC-2 using, say,
10.0.0.1 cannot be uniquely associated to a single destination VPC. At
least, not without a flow session table lookup.

In preparation for this lookup, write the destination VPC associated to
a flow to the value for this flow's entry in the flow table. In
follow-up commits, we'll use it to support shared IP addresses between
different VPCs.

Signed-off-by: Quentin Monnet <qmo@qmon.net>