test(nat): Extend test for shared IPs with stateful NAT
Now that we support "shared IPs" with stateful NAT, that is, overlapping
IPs from different Expose blocks, which makes a flow table lookup
necessary to determine the destination VPC for some packets.

One case is when multiple VPCs expose overlapping CIDR ranges to a given
VPC: in that case, the destination VPC for return traffic from this
given VPC cannot be determined based on the source VPC and destination
address, only. We already have a test for this situation: this commit
extends it to make sure that the lookups updated in recent changes, and
the NAT transformation, work as expected.

We also add a test to cover the case when a pair of VPCs has multiple
peerings that expose overlapping CIDR ranges. This case is not reachable
in practice today, because we reject such configurations when validating
the configuration, but we can run the test anyway.

Signed-off-by: Quentin Monnet <qmo@qmon.net>