fix(nat): Fix timeout reset for table flows in stateful NAT
When updating the timeout for a flow after a successful lookup in the
flow table, we add the desired timeout duration to the existing timeout
instant, instead of the current time, meaning that we linearly increment
the duration (beyond the desired period) each time we receive a packet
for that flow. Let's fix by resetting to the current time, plus the
desired duration, instead.

We do this by introducing a new method reset_expiry(), doing

    self.expires_at.store(
        Instant::now() + duration,
        std::sync::atomic::Ordering::Relaxed,
    );

To compare with extend_expiry(), which does:

    self.expires_at
        .fetch_add(duration, std::sync::atomic::Ordering::Relaxed);

Fixes: 9d9e22906189 ("fix(nat): Fix timeout for flow expiration")
Reported-by: Fredi Raspall <fredi@githedgehog.com>
Signed-off-by: Quentin Monnet <qmo@qmon.net>