feat(nat): Add 3rd stateful NAT session entry for destination VPC lookup
We want to support shared IP addresses exposed from VPCs, so that both
VPC-1 and VPC-3 can expose the same CIDR to VPC-2. The usual destination
VPC lookup is insufficient in that case, we have to resort to a flow
session lookup. In a previous commit, we made the NAT stage write the
destination VPC discriminant to flow entries' values.

This is not enough: we also need to adjust the way we do the lookup.
Specifically, there are cases when we don't know what the destination
VPC discriminant is: this is why we added the discriminant to the flow
value, so we could retrieve it. In that case, we cannot rely on the
destination discriminant as part of the flow key, precisely because we
ignore it. For such cases, we need an entry with an empty discriminant
in the key for the lookup,

To get this entry, we create a third session entry for new sessions. We
already had two entries: one for the forward path, one for the reverse
path. We add a third one, for the reverse path, but with a null
destination VPC discriminant in the key, so that we can match against it
for return flows, when we are not able to figure the destination VPC
from the dedicated lookup stage.

We could optimise the creation of these new entries if we had an easy
way to tell, at session creation time, what connections are susceptible
to "collide" at the destination lookup stage. We could generate a list
of these at configuration time, to avoid adding the third flow table
entry when it is unnecessary, but we don't have this implemented at the
moment.

Note: A previous attempt at addressing the lookup for shared IPs relied
on zero-ing the destination VPC discriminant for all keys at session
creation and at lookups, but this turned out to not be viable because
some sessions might collide in the flow table.

Suggested-by: Manish Vachharajani <manish@githedgehog.com>
Signed-off-by: Quentin Monnet <qmo@qmon.net>