chore(dataplane): Remove disgusting force_send hack
tokio::task::JoinSet::spawn requires the passed future to be
Send, but our pipeline is not Send.  force_send was a hack to
make it Send in the type system since we knew up front that
we are in a single thread runtime.

Good news!, tokio JoinSet has a method called spawn_local that
keeps the spawned task on the local thread and does not require
the future to be Send.

This commit uses spawn_local and removes force_send.

Signed-off-by: Manish Vachharajani <manish@githedgehog.com>