ci: Add step to print images name and tags
Add a quick step to development CI workflow to print the names and tags
of pushed container images. These names are already available from the
previous step ("push container"), but it might be tedious to look for
them in the logs. This additional step makes it easier to retrieve the
tags.

Signed-off-by: Quentin Monnet <qmo@qmon.net>