feat(k8s-intf): Generate Rust types for gateway-agent CRD
This is the first step to getting rid of the gateway agent
and having dataplane talk directly to k8s.  This should
greatly simplify the code, eliminate the gRPC conversion
and eliminate the entire gateway-proto repository.

This commit adds the k8s-intf crate with build
infrastructure that uses kopium to generate the Rust types
for the CRD for the gateway-agent that is in the gateway
repository.

The next steps will use the CRD to fetch and parse data
from k8s, and then translate it into the inner config
format of the dataplane.

Signed-off-by: Manish Vachharajani <manish@githedgehog.com>