test(nat): Allow turning off random port allocation (stateful NAT tests)
Add a variant of the method to update an allocator, only compiled for
tests, so that we can turn off randomness for port block ordering for
stateful NAT port allocation in tests. This matters, for example, when
we want to check whether, all other items being equal, we can have
collisions on allocated ports, and hence on allocated NAT sessions, in
specific test cases.

Signed-off-by: Quentin Monnet <qmo@qmon.net>