test(nat): Update NAT test to show existing issue with shared IPs
We recently added support for "shared IPs" in stateful NAT (for example,
when two different VPCs expose the same NAT-ed IPs to a given VPC), but
the design has an issue: when the exposed IPs on both sides of the two
peerings of our example are identical, we risk creating identical flow
table entries for the destination VPC lookup stage (the keys for the
entries for reverse connections from the VPC seeing shared IPs would
normally differ on their destination VPC discriminant, but this field is
set to None for the entries for destination VPC lookup, precisely
because we don't know them at the lookup stage).

Update our test to illustrate this scenario.

Why didn't the test catch it sooner? Two reasons:

- We had "random" port allocation turned on, meaning that collisions on
  the source port for packets from VPC-1 to VPC-2 and VPC-3 to VPC-2
  were unlikely in the first place - although I did try turning it off
  manually in the code before merging the test.

- The main problem is that we first try to see if a packet from VPC-3
  with different IPs and ports from the one sent from VPC-1 triggers an
  issue. It doesn't, of course, the flow table entries are different.
  But this packet "picks" the first available port from the port range
  for single allocated IP address, which means that when we later try
  with a packet from VPC-3 with the same IPs and ports as the packets
  from VPC-1, we cannot get the same port allocated (we get the second
  of a port block, while the packets from VPC-1 use the first port of
  their port block).

To make sure we trigger the issue, turn off "random" port allocation,
and remove the intermediate step that used different IPs and ports.

Make sure to document the issue so we can come back and adjust the test
once the issue is fixed.

The issue itself is tracked at
https://github.com/githedgehog/dataplane/issues/1083.

Signed-off-by: Quentin Monnet <qmo@qmon.net>