refactor(nat): Simplify stateless NAT code logic
Turns out a good portion of the code can be simplified, and made
clearer, in the main stateless NAT logic.

We also move some helper functions, addr_offset_in_range(),
addr_from_offset(), and map_ip_nat(), to the table module: they're not
directly related to table manipulation, but they depend entirely depend
on how the stateless NAT tables are implemented (getting IP offsets in
table entries depends on the structure of these entries). This way, we
query for the direct mapping from the main module, without working with
the notion of "ranges" and offset which is proper to the table's
internals.

Signed-off-by: Quentin Monnet <qmo@qmon.net>