feat(nat): Switch to LPM trie for stateless NAT tables
To store the context for stateful NAT translation, we've been using a
BTreeSet with IP addresses representing the start of initial ranges as
keys, and the end of the corresponding initial range, plus the start of
the associated translated range, as values. This allowed us to perform a
lookup by searching the tree for the nearest lower range bound.

Would a LPM structure be more efficient instead? We've had some
discussions about that. Given that the list of initial prefix has no
overlap, there's no notion of "longest" prefix match for stateful
address translation, there's just one prefix, at most, that a given
address can match. So the LPM loses some of its advantages. Furthermore,
the BTreeSet we've used allowed us to flatten and regroup, when
possible, address ranges, regardless of the CIDR format, so we could
have a direct range-to-range mapping, without ever having to iterate
over several potential ranges associated with a given prefix.

Anyway.

We're now looking at adding support for port translation as well, as
part of stateless NAT. If we map, say, 1.0.0.0/16 with a given port
range 1000-2000, and 1.0.1.0/24 with a distinct port range 5000-6000,
then we do have overlap for the IP prefixes, and a LPM trie makes more
sense. It also makes it easier to understand the structure and to extend
it. Because of the prefix overlap, it would also become unfeasible in
practice to store the ranges in the list in such a way that we keep them
ordered: we'd need to insert each IP and its port in order:
1.0.1.0:1000-2000, then 1.0.1.0:5000-6000, then 1.0.1.1:1000-2000, then
1.0.1.1:5000:6000, each with its own corresponding range. We're much
better with a LPM trie.

This commit switches the stateless NAT tables implementation to a LPM
trie, using our existing implementation of an IpPrefixTrie, and adjusts
the tests accordingly.

When building the tables, we no longer try to merge contiguous
prefixes or matching ranges together. Prefix merging, when relevant,
should be done at config processing time. Merging ranges will be added
as a follow-up commit.

Signed-off-by: Quentin Monnet <qmo@qmon.net>