feat(nat): Restore range merge when building stateless NAT context
In the LPM trie that we now use for stateless NAT tables, each initial
prefix gets associated with a vector of target ranges. This target
ranges are disjoint; but then can be contiguous, in which case, it's
interesting to merge them together, so that we get fewer ranges to
process when mapping a given address - see the updated test for
examples. We lost target range merging when moving the stateless NAT
tables to the new LPM-based implementation; this commit restores it. We
also lost merging for initial ranges, but given that we now use CIDR
prefixes as keys, we don't have as much flexibility as before to merge
them (we can only merge CIDR prefixes when it creates a new valid CIDR,
and this is best left to the config processing code).

Signed-off-by: Quentin Monnet <qmo@qmon.net>