feat(nat): do not consider src vpc in PoolTableKey
When using stateful NAT with addresses shared by distinct VPCs,
the source port (translated) must be such that the return traffic
allows to unambiguously recover the original flow. The existing
implementation considers the source vpc id as part of the key to
locate a port allocator. This has the problem that two flows
from distinct vpcs natting to the same src IP may choose the same
port since distinct port allocators are used. Fix this by not
considering the source vpc id as part of the key to select a port
allocator. This fixes the problem in general, without breaking
the behavior when nat addresses are not reused across vpcs.

When NAT addresses are re-used among VPCs, it may happen that a
VPC uses many more ports than the rest, potentially starving the
latter. This can and should be addressed later by tracking the
number of sessions/ports-allocated on a per VPC level wrt the
fair share of ports, which should be the number of ports available
over the number of vpcs. This is left for future work, and
classical resource-sharing policies may be applied like having
a strict quota or a relaxed one with possible eviction/pre-emption.

However, before addressing the above, the number of available
sessions per nat IP address could be significantly increased by
changing the criteria to allocate ports: ideally, with stateful
source NAT, a distinct port allocator should be used per
(nat_src_ip, dst_ip, protocol, dst_port and dst_vpc_id).
Currently the dst_ip and dst_port are not taken into account.

Signed-off-by: Quentin Monnet <qmo@qmon.net>
Signed-off-by: Fredi Raspall <fredi@githedgehog.com>