fix(k8s-intf): Convert i32 to u32 in kopium generated code
kopium generates types from the openapi schema that is part
of k8s CRD definitions.  Unfortunately, JSON and openapi
only have an integer type that is signed.  As a result, the
k8s tools for Go will take uint32 from a Go struct and
generate a CRD where the openapi type for the field is
integer and where format is int32, because openapi format
hints also do not have unsigned types.

Then, kopium looks at this CRD and generates an i32 in rust
for that field.  But this is a problem because the actual
value in JSON can be larger than what will fit into an i32
despite the openapi format hint.  kube-rs will then use
serde_json to parse the json and the rust side will fail
to parse json outputted by go if the uint32 in go will not
fit in a rust i32.

The cleanest fix from an API POV for this is to convert the
generated types for known problem fields from i32 to u32
so that all values will be parsed correctly.  Sadly, the
implementation of this rewrite is simple but gross.  See
build.rs in this commit for the implementation.

Signed-off-by: Manish Vachharajani <manish@githedgehog.com>