feat(config): Add convert k8s CRD to Underlay config
This commit adds conversion for
k8s_intf::gateway_agent_crd::GatewayAgentGateway
into a proper Underlay configuration.

Bi-directional conversion is provided for interfaces
and neighbors where there is a good mapping between
the Underlay types and the CRD types.

For the default VRF, where there is no corresponding
type for the CRD, the default VRF is generated
explicitly.

Signed-off-by: Manish Vachharajani <manish@githedgehog.com>