chore(k8s-intf): Cleanup empty collection to None logic
Move from using if c.is_empty() { None } else { Some(c) }
which falls across multiple lines and results in uglier
construction of objects.

This commit uses Some(c).filter(|c| !c.is_empty()) instead
as this is a one liner and formats nicely, even though it
may be a bit slower because Some(c) is constructed.

In the case of this commit, Some(c.normalize()) is called
which is even more inefficient, but for the purposes of the
modified test code, the better formatting is preferable to
the slightly enhanced performance.

Signed-off-by: Manish Vachharajani <manish@githedgehog.com>