chore: permit RUSTSEC-2025-0134
RUSTSEC-2025-0134 is a very minor issue which does not impact us in
any case.

All that has actually happened is that a thin wrapper around rustls-pki-types
(i.e. rustls-pemfile) was archived (the main repo is still maintained).

This is irrelevant to us on multiple levels:

1. this is only indirectly pulled in via testn,
2. in dev dependencies,
3. for a feature we disable explicitly.

I created an issue in the bollard repo to address this

<https://github.com/fussybeaver/bollard/issues/612>

Signed-off-by: Daniel Noland <daniel@githedgehog.com>