build(hardware): remove unused dependency on sysroot
Removing this build.rs file facilitates removing dpdk as a strict requirement for the build.

The build.rs file for the hardware package is not strictly needed. Strictly speaking, the only packages which actually
need this type of build.rs logic are those which produce an actual linked artifact (.so or elf exe or some such).

Signed-off-by: Daniel Noland <daniel@githedgehog.com>