build(init): make sysroot optional
The sysroot is likely functionally necessary for most developers to build the init package (due to the dependency on the
hardware crate), it is still possible that the developer will have their own version of the required libraries
available.
Such builds are entirely unsanitary and can never be used in production, but my be helpful during development.

Signed-off-by: Daniel Noland <daniel@githedgehog.com>