build: drop most of config.toml
This change is temporary and actually reflects substantial attrition in our build process.
Specifically, this change exposes us to various forms of sterility violations for local development builds.

That said, CI should still be sterile since the versions of rustc, clant, and all sysroot libs shipped in the sterile
env are also the one and only valid choice to compile with / link against in those contexts.
This step is also an essential proof of concept validation for the removal of the sysroot as a hard dependency.

In the very near future we should restructure this project to use nix flakes to avoid the potential for the ever popular
"works on my machine."
We haven't had a lot of "works on my machine" in the project to date, and I would very much like to keep it that way.

Signed-off-by: Daniel Noland <daniel@githedgehog.com>