feat: Add port range to struct VpcExpose
Prepare for Port Address Translation in NAT. This commit is based on a
simple change. The following struct:

     pub struct VpcExpose {
         pub ips: BTreeSet<Prefix>,
         pub nots: BTreeSet<Prefix>,
         pub nat: Option<VpcExposeNat>,
     }

becomes:

     pub struct VpcExpose {
         pub ips: BTreeSet<PrefixWithOptionalPorts>,
         pub nots: BTreeSet<PrefixWithOptionalPorts>,
         pub nat: Option<VpcExposeNat>,
     }

In other words, we attach an optional port range to each prefix in the
various prefix lists of a VpcExpose object. Such objects are used in
several locations in the code, and as a consequence, we need to update
all of these. This includes:

- The gRPC/K8s conversion methods in crate config
- The validation steps for VpcExpose objects in crate config
- The "prefix collapse" utilities in crate config
- Building route advertisement in crate mgmt
- The stateful NAT allocator in crate nat
- The stateless NAT context build in crate nat
- The destination VPC lookup stage in crate pkt-meta

WARNING: These locations get trivial updates to make the code compile.
The surrounding logic is not updated to account for port ranges in this
commit. It will be updated, where relevant, in follow-up commits.

Signed-off-by: Quentin Monnet <qmo@qmon.net>