feat(lpm): Introduce DisjointRangesBTreeMap
Introduce a new struct DisjointRangesBTreeMap, containing a generic
BTreeMap.

This struct is meant to be used with map keys representing
non-overlapping ranges, for example: disjoint port ranges or disjoint IP
prefixes.

The idea is the following: we can store IP ranges, for example, in a
DisjointRangesBTreeMap object, and look up for a single IP address, to
find the corresponding map key (the IP range) and the associated value,
whatever it is. The struct implements this lookup, so that we don't have
to re-implement it each time we want to use a similar BTreeMap with
different element types. The way we perform this lookup is by retrieving
the closest lower map key (IP range) matching the IP address, using
methods provided by the BTreeMap.

We already use similar structs in several locations in the code, and we
used to have it for stateless NAT in the past (before we switched to a
LPM). We also plan on adding a new one for the destination VPC lookup
discriminant stage, in a future commit. Introducing a
DisjointRangesBTreeMap should help harmonise these, although we don't
convert the existing similar structs at the moment.

Notes:

- As for the location of this struct: should it be in crate lpm? It's
  not directly related to the LPM implementation we have, but I didn't
  find a location that would really be better suited, and I don't think
  it deserves a dedicated crate. The crate also hosts the implementation
  for the Prefix structs, which are not completely tied to the LPM,
  either.

- Maybe re-implementing many methods (iter(), range(), ...) for the
  DisjointRangesBTreeMap, just piggybacking on the methods for the inner
  BTreeMap, might be overly verbose. If it's too much we may change in
  the future to expose the inner BTreeMap, and use the methods from
  there, although I'd rather keep it encapsulated if possible.

Signed-off-by: Quentin Monnet <qmo@qmon.net>