feat(pkt-meta): Adjust destination VPC lookup for port ranges
In preparation for supporting port ranges in "expose" objects, for Port
Address Translation, adjust the destination VPC discriminant lookup
stage.

So far, for a given packet, the source VPC and the destination VPC would
usually provide the destination VPC. When we add port ranges to
VpcExpose objects, we may have two packets from the same VPC, with the
same destination IP, but with L4 destination port from different ranges,
maybe associated to distinct peerings corresponding with distinct VPCs.
In this case, the source VPC and destination address are no longer
enough to determine the destination VPC: we must take the destination
port into account, too. This commit adjust the VpcDiscriminantTable and
related lookup accordingly.

As part of this, we re-implement the lookup() and insert() method for
the VpcDiscriminantTable itself, in order to mask the additional
complexity when working with the inner IpPrefixTrie.

Signed-off-by: Quentin Monnet <qmo@qmon.net>