feat(config,lpm): Adjust prefix collapsing to account for port ranges
"Collapsing" prefixes, in the context of this commit, comes down to
"applying" exclusion prefixes to a set of prefixes, by splitting the
latter set into subsets containing all the initial addresses, minus
those from the exclusion prefixes. This is helpful for NAT, because
after we've collapsed prefixes, we don't need to work with exclusion
prefixes anymore.

In preparation for Port Address Translation, we've updated VpcExpose to
associate optional port ranges to IP prefixes. As a result, we now need
to account for port ranges when "applying" exclusion
prefixes-with-port-ranges to prefixes-with-port-ranges. This is not
trivial to do, but with a bit of refactoring, it is not as scary as it
sounds.

Signed-off-by: Quentin Monnet <qmo@qmon.net>