refactor(nat): Rename TrieRange as IpRange, ip_len() as len()
This is some simple renaming, in preparation for more changes in the
code for the stateless NAT tables.

Struct TrieRange will get an alternative struct with port ranges, soon.
Make sure we can distinguish the two by renaming the existing one as
what is is: an IP address range.

Once we've renamed it, it no longer makes sense to call it's length
method ip_len(). Instead, just make it len().

Signed-off-by: Quentin Monnet <qmo@qmon.net>