feat(dataplane): set gateway name
Set the gateway name from the cmd line --name argument.
If no --name is provided, set the gateway name from the hostname
provided by the OS. Fail to start if the gatway name cannot be set.

With this, the gateway name defaults to the hostname but can always
be overriden with --name.

Signed-off-by: Fredi Raspall <fredi@githedgehog.com>