feat(nat): Update structure of stateless NAT tables for port ranges
Working towards support for Port Address Translation in stateless NAT,
update the structure of the stateless NAT tables to account for port
ranges: we turn NatTableValue into an enum with variants
AddrTranslationValue (IP range only) and PortAddrTranslationValue (IP
and port ranges). We introduce the IpPortRange type, similar to IpRange
but with an associated port range. It differs from PrefixWithPorts from
lpm crate in that the IP range may not be a CIDR prefix.

As for the map lookup, we used to do a simple LPM lookup to find the IP
range (LPM key) corresponding to a given IP address, we'd retrieve
associated target prefixes (LPM value) and do the mapping. IP ranges in
the LPM never overlapped. With ports, it's a bit different: we can have
overlapping IP ranges (as long as their respective associated port
ranges make them disjoint). As a consequence, the longest matching IP
prefix may no longer be the one matching the IP _and the port_ from the
packet. Instead, we need to iterate over _all matching IP prefixes_ to
find the one with the correct associated port. See method
matching_entries() added to the LPM.

Adjustments for building the table will come in a follow-up commit.

Signed-off-by: Quentin Monnet <qmo@qmon.net>